🚧 (wip) extract state from stacked discrete bar charts
